+++ NEWS.md
@@ -1,3 +1,5 @@
+# selenium (development version)
+
 # selenium 0.1.3
 
 * The `browser`, `host` and `port` fields can now be used to access the browser,