+++ NEWS.md
@@ -1,4 +1,4 @@
-# selenium (development version)
+# selenium 0.1.4
 
 - Added `chrome_options()`, `firefox_options()` and `edge_options()` to help
   with the `capabilities` argument in `SeleniumSession$new()`. The documentation